Add files via upload
Enclosed are more recent scripts for Warlike. In particular, warliketurfwarspawnscript and warlikespawnscript2 are in their most recent form here. They look like they were decompiled with Champollion because they were; I couldn't find the originals and had to resort to decompiling them from the .BA2 through Nexus in order to make changes and update the mod properly. This happened because I used Warlike's "engine" to build The Boston Massacre, so many of the scripts were altered and would take too long to revert.